aria-backend/requirements.txt
google-adk>=1.25.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.18.0
python-dotenv>=1.0.0
playwright>=1.50.0
